Cooling bug and error fixes
A cooling bug was reported by Jasper, where a pause did not shut down the cooling inlet. Setting the temperatures to 0 did not clear the cooling flag, which resulted in attempting to cool down to 0C. This flag is cleared now, which should prevent this from happening. Many people reported getting heater errors when the heaters were working fine. I discovered a small bug in the math, which I fixed, but I also added a minimum on time for the heaters so that the error check doesn't happen for very short heater pulses, which I believe was the main cause.